requirements.txt
#Core libraries to install
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.5
Pillow>=8.0.0
torch>=1.10.0
tensorflow>=2.6.0
keras>=2.6.0
scikit-learn>=0.24.0
ultralytics>=8.3.67 # YOLO models
ultralytics-thop==2.0.14
python-dotenv~=1.0.1
